Updated query to get redemptions

+++ src/main/java/com/shakepoint/web/api/core/repository/impl/PromoCodeRepositoryImpl.java
@@ -72,8 +72,8 @@ public class PromoCodeRepositoryImpl implements PromoCodeRepository {
 
     @Override
     public Integer getPromoCodeRedemptions(String code) {
-        return (Integer) em.createQuery("SELECT COUNT(r.id) FROM PromoRedeem r WHERE r.promoCode.code = :code")
-                .setParameter("code", code).getSingleResult();
+        return ((Long) em.createQuery("SELECT COUNT(r.id) FROM PromoRedeem r WHERE r.promoCode.code = :code")
+                .setParameter("code", code).getSingleResult()).intValue();
     }
 
     @Override
